+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/plugin_list.rb
@@ -17,7 +17,7 @@ module AwsSdkCodeGenerator
 
     def compute_plugins(options)
       plugins = {}
-      plugins.update(default_plugins)
+      plugins.update(options[:async_client] ? default_async_plugins : default_plugins)
       plugins.update(signature_plugins(options.fetch(:signature_version)))
       plugins.update(protocol_plugins(options.fetch(:protocol)))
       plugins.update(options.fetch(:add_plugins))
@@ -55,6 +55,17 @@ module AwsSdkCodeGenerator
         'Aws::Plugins::ClientMetricsPlugin' => "#{core_plugins}/client_metrics_plugin.rb",
         'Aws::Plugins::ClientMetricsSendPlugin' => "#{core_plugins}/client_metrics_send_plugin.rb",
       }
+    end
+
+    def default_async_plugins
+      plugins = default_plugins.dup
+      plugins.delete('Aws::Plugins::ResponsePaging')
+      plugins.delete('Aws::Plugins::EndpointDiscovery')
+      plugins.delete('Aws::Plugins::EndpointPattern')
+      plugins.delete('Aws::Plugins::ClientMetricsPlugin')
+      plugins.delete('Aws::Plugins::ClientMetricsSendPlugin')
+      plugins['Aws::Plugins::InvocationId'] = "#{core_plugins}/invocation_id.rb"
+      plugins
     end
 
     def protocol_plugins(protocol)