+++ gems/aws-sdk-core/lib/aws-sdk-core.rb
@@ -19,6 +19,7 @@ require_relative 'aws-sdk-core/process_credentials'
 # client modules
 
 require_relative 'aws-sdk-core/client_stubs'
+require_relative 'aws-sdk-core/async_client_stubs'
 require_relative 'aws-sdk-core/eager_loader'
 require_relative 'aws-sdk-core/errors'
 require_relative 'aws-sdk-core/pageable_response'