+++ gems/aws-sdk-core/lib/aws-sdk-core/binary/event_stream_decoder.rb
@@ -7,11 +7,13 @@ module Aws
 
       # @param [String] protocol
       # @param [ShapeRef] rules ShapeRef of the eventstream member
+      # @param [ShapeRef] output_ref ShapeRef of output shape
+      # @param [Array] error_refs array of ShapeRefs for errors
       # @param [EventStream|nil] event_stream_handler A Service EventStream object
       # that registered with callbacks for processing events when they arrive
-      def initialize(protocol, rules, io, event_stream_handler = nil)
+      def initialize(protocol, rules, output_ref, error_refs, io, event_stream_handler = nil)
         @decoder = Aws::EventStream::Decoder.new
-        @event_parser = EventParser.new(parser_class(protocol), rules)
+        @event_parser = EventParser.new(parser_class(protocol), rules, error_refs, output_ref)
         @stream_class = extract_stream_class(rules.shape.struct_class)
         @emitter = event_stream_handler.event_emitter
         @events = []
@@ -42,6 +44,7 @@ module Aws
         case protocol
         when 'rest-xml' then Aws::Xml::Parser
         when 'rest-json' then Aws::Json::Parser
+        when 'json' then Aws::Json::Parser
         else raise "unsupported protocol #{protocol} for event stream"
         end
       end