+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Support HTTP/2 based AWS event stream operations
+
 3.46.2 (2019-02-19)
 ------------------